Organize loaders module
- Move loaders GraphQL stuff to schema.ts

- Group loaders into submodules:
    - `loaders/schema`: for the ArtifactsLoader and GraphQL hookup
    - `loaders/resources`: for bytecodes/compilations/etc. to generate
      the graphql mutations, variables

- Rename AddCompilation -> AddCompilations

- Move all queries modules to respective resource module

- Initialize empty resource modules as a home for the GraphQL
  mutation, and later to house the actual loading logic.

- Align all GraphQL mutation modules (was: loaders/queries/*.ts) to
  match form `loaders/resources/<resource>/add.graphql.ts`
