+++ packages/truffle-db/src/loaders/resources/bytecodes/add.graphql.ts
