+++ packages/truffle-db/src/loaders/resources/contractInstances/add.graphql.ts
