+++ packages/truffle-db/src/loaders/resources/contracts/add.graphql.ts
