+++ packages/truffle-db/src/loaders/resources/contracts/index.ts
@@ -1,0 +1,2 @@
+import { AddContracts } from "./add.graphql";
+export { AddContracts };
