+++ packages/truffle-db/src/loaders/resources/networks/add.graphql.ts
