+++ packages/truffle-db/src/loaders/resources/networks/index.ts
@@ -1,0 +1,1 @@
+export { AddNetworks } from "./add.graphql";
