+++ packages/truffle-db/src/loaders/schema/test/compilationSources/SquareLib.sol
