+++ packages/truffle-db/src/loaders/schema/test/index.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import path from "path";
 import gql from "graphql-tag";
 import { TruffleDB } from "truffle-db";
-import { ArtifactsLoader } from "truffle-db/loaders/artifacts";
+import { ArtifactsLoader } from "truffle-db/loaders/schema";
 import { generateId } from "truffle-db/helpers";
 import * as Contracts from "@truffle/workflow-compile/new";
 import Migrate from "@truffle/migrate";
