+++ packages/truffle-db/src/loaders/schema/test/workflowCompileOutputMock/compilationOutput.json
